fix intersection ValueError
checks if intersection_segment_plane returns the mesh vertex (u or v of the edge ). 
if True appends the mesh vertex itself to intersections instead of calculating t. otherwise when mesh.split_edge is executed, it raises ValueError: t should be greater than 0.0. or ValueError: t should be smaller than 1.0.